requirements.txt
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
PyYAML==6.0.2
PyYAML==6.0.3
scikit_learn==1.7.2
torch==2.5.1
tqdm==4.67.1
umap==0.1.1
